backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-pptx==0.6.21
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
Pillow==10.0.1
langchain==0.0.335
openai==1.3.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.7